SC 110072: Separate batched component files into individual components
