+++ src/applications/appeals/995/components/EvidenceVaRecords.jsx
@@ -16,11 +16,9 @@ import {
 } from '../validations/evidence';
 
 import { focusEvidence } from '../../shared/utils/focus';
-import {
-  HeaderAndModal,
-  IssueAndDates,
-  PageNavigation,
-} from './EvidenceRecords';
+import { EvidenceHeaderAndModal } from './EvidenceHeaderAndModal';
+import { IssueAndDates } from './IssueAndDates';
+import { PageNavigation } from './PageNavigation';
 
 import { getIssueName, getSelected } from '../../shared/utils/issues';
 import { checkValidations } from '../../shared/validations';
@@ -316,7 +314,7 @@ const EvidenceVaRecords = ({
   return (
     <form onSubmit={handlers.onGoForward}>
       <fieldset>
-        <HeaderAndModal
+        <EvidenceHeaderAndModal
           currentData={currentData}
           currentState={currentState}
           currentIndex={currentIndex}
